Parallel speedup works again, on Savio2 nodes at least. Tested whole chromosome hdf5 output, merging is much faster than for bigwig. Memory usage controls work but where exactly memory is being used remains unclear: memory profiling will likely be valuable.

+++ dimelo/utils/read_parser.py
@@ -263,21 +263,21 @@ def parse_read_by_basemod(
     # Build output arrays
     ######################################################################################## 
 
-    reference_coordinates = np.arange(read_start,read_end)
+    reference_coordinates = np.arange(read_start,read_end,dtype=int)
     
     # Positions where the read aligns, so we can get read depth
-    read_coordinates = np.zeros(read_end-read_start)
+    read_coordinates = np.zeros(read_end-read_start,dtype=int)
     if len(read_indices)>0:
         read_coordinates[np.array(reference_positions_rel)[np.array(read_indices)].astype(int)] = 1
     
     # Positions where the context is valid for a base modification, so we can get pileup denominator
-    valid_coordinates = np.zeros(read_end-read_start)
+    valid_coordinates = np.zeros(read_end-read_start,dtype=int)
     if len(valid_indices)>0:
         valid_coordinates[np.array(reference_positions_rel)[np.array(valid_indices)].astype(int)] = 1
     
     # Positions where the base-in-context was actually modified, so we can get pileup numerator. 
     # Report the actual prob if threshold was zero
-    modified_coordinates = np.zeros(read_end-read_start)
+    modified_coordinates = np.zeros(read_end-read_start,dtype=float)
     if len(modified_indices)>0:
         if threshold>0:
             modified_coordinates[np.array(reference_positions_rel)[np.array(modified_indices)].astype(int)] = 1
